Rename vulnerability-formatter
