+++ packages/dd-trace/src/appsec/iast/vulnerability-reporter.js
@@ -1,6 +1,6 @@
 const { MANUAL_KEEP } = require('../../../../../ext/tags')
 const LRU = require('lru-cache')
-const vulnerabilitiesFormatter = require('./vulnerabilities-formatter/vulnerability-formatter')
+const vulnerabilitiesFormatter = require('./vulnerabilities-formatter')
 const VULNERABILITIES_KEY = 'vulnerabilities'
 const IAST_JSON_TAG_KEY = '_dd.iast.json'
 const VULNERABILITY_HASHES_MAX_SIZE = 1000
