+++ packages/dd-trace/test/appsec/iast/vulnerability-formatter/vulnerability-formatter.spec.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const vulnerabilityFormatter = require('../../../../src/appsec/iast/vulnerabilities-formatter/vulnerability-formatter')
+const vulnerabilityFormatter = require('../../../../src/appsec/iast/vulnerabilities-formatter')
 
 const { suite } = require('./resources/evidence-redaction-suite.json')
