+++ packages/dd-trace/test/appsec/iast/vulnerability-reporter.spec.js
@@ -493,7 +493,7 @@ describe('vulnerability-reporter', () => {
   })
 
   describe('evidence redaction', () => {
-    const vulnerabilityFormatter = require('../../../src/appsec/iast/vulnerabilities-formatter/vulnerability-formatter')
+    const vulnerabilityFormatter = require('../../../src/appsec/iast/vulnerabilities-formatter')
 
     beforeEach(() => {
       sinon.spy(vulnerabilityFormatter, 'setRedactVulnerabilities')
